feat: nestjs implementation
Removes tests and sets everything in a single file.
